refactor: move prnt_scrn to separate module
