+++ index.js
@@ -1,9 +1,9 @@
-import Jimp from 'jimp';
 import { httpServer } from './src/http_server/index.js';
 import robot from 'robotjs';
 import { WebSocketServer, WebSocket, createWebSocketStream } from 'ws';
 import { drawCircle } from './src/draw-circle.js';
 import { drawRectangle } from './src/draw-rectangle.js';
+import { prntScrn } from './src/prnt-scrn.js'
 
 const HTTP_PORT = 3000;
 
@@ -25,33 +25,30 @@ function onConection(ws) {
 async function onMessage(message) {
     const command = message.toString('utf8').split(' ')[0];
     let params = message.toString('utf8').split(' ').slice(1).map(item => Number(item));
-    robot.setMouseDelay(5);
+    robot.setMouseDelay(10);
     const mousePos = robot.getMousePos();
     switch (command) {
         case 'mouse_left':
-            robot.moveMouseSmooth(mousePos.x - params[0], mousePos.y);
+            robot.moveMouse(mousePos.x - params[0], mousePos.y);
             break;
         case 'mouse_right':
-            robot.moveMouseSmooth(mousePos.x + params[0], mousePos.y);
+            robot.moveMouse(mousePos.x + params[0], mousePos.y);
             break;
         case 'mouse_up':
-            robot.moveMouseSmooth(mousePos.x, mousePos.y - params[0]);
+            robot.moveMouse(mousePos.x, mousePos.y - params[0]);
             break;
         case 'mouse_down':
-            robot.moveMouseSmooth(mousePos.x, mousePos.y + params[0]);
+            robot.moveMouse(mousePos.x, mousePos.y + params[0]);
             break;
         case 'draw_square':
-            console.log('draw_square');
             const side = params[0];
             drawRectangle(robot, side, side);
             break;
         case 'draw_circle':
             const r = params[0];
             drawCircle(robot, r);
-
             break;
         case 'draw_rectangle':
-            console.log('draw_rectangle');
             const sideX = params[0];
             const sideY = params[1];
             drawRectangle(robot, sideX, sideY);
@@ -64,37 +61,17 @@ async function onMessage(message) {
                     messageStream = createWebSocketStream(client, { decodeStrings: false, encoding: 'utf8' });
                     messageStream.write(coordinates);
                 }
-
             })
             break;
 
         case 'prnt_scrn':
-            const size = 200;
-
-            const img = robot.screen.capture(mousePos.x - size / 2, mousePos.y - size / 2, size, size);
-            const data = [];
-            const bitmap = img.image;
-            let i = 0, l = bitmap.length;
-            for (i = 0; i < l; i += 4) {
-                data.push(bitmap[i + 2], bitmap[i + 1], bitmap[i], bitmap[i + 3]);
-            }
-
-            const jImg = new Jimp({
-                "data": new Uint8Array(data),
-                "width": img.width,
-                "height": img.height
-            });
-
-            const imgString = await jImg.getBase64Async(Jimp.AUTO);
-            var buf = `prnt_scrn ${imgString.substring(22)}`;
-
+            const imgString = await prntScrn(robot);
             server.clients.forEach(client => {
                 if (client.readyState === WebSocket.OPEN) {
                     messageStream = createWebSocketStream(client, { decodeStrings: false });
-                    messageStream.write(buf);
+                    messageStream.write(imgString);
                 }
             })
-
             break;
         default:
             break;
